NLP/requirements.txt
paralleldots
keras
pytorch
torchvision
tensorflow-gpu
matplotlib
scikit-learn
nltk
matplotlib
seaborn
pandas
jupyter
tqdm
xlrd
jieba
tinysegmenter
cltk